requirements.txt
aws-cdk-lib==2.127.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment==3.0.33
aws-cdk.lambda-layer-kubectl-v27
